refactor(explore): Move mock associations to a different file and refactor explore screen.
- Moved the mock data type the list to a different file (in model/associations).
- Refactored some functional lines of code for better readability
